Sales by Match Hacker Rank Code
Where the sales person needs to sell the scoks , but before that he must select or segregate the socks into pairs and then count on it.

Here the collections python module is used to calculate the number of times the value or number appears and it would return the dictionary. Then a for loop to count it.